Introduce a cloud-init-aws-slave script
This is necessary for new slave instances starting on AWS
as their storage drive is ephemeral, therefore when new
instances are booted we still need to format their NVMe
drive before using it.

In addition, we need some directories to be created on
this drive to allow permissions to the Jenkins agent.